refactor(anywidget): use signals for runtime
