+++ packages/anywidget/package.json
@@ -22,7 +22,8 @@
 	"dependencies": {
 		"@anywidget/types": "workspace:~",
 		"@anywidget/vite": "workspace:~",
-		"@jupyter-widgets/base": "^2 || ^3 || ^4 || ^5 || ^6"
+		"@jupyter-widgets/base": "^2 || ^3 || ^4 || ^5 || ^6",
+		"solid-js": "^1.8.14"
 	},
 	"devDependencies": {
 		"@jupyter-widgets/base-manager": "^1.0.6",
